wip: move filter/sort logic from useGithub hook to the main app
